backend/requirements.txt
flask
pytesseract
pillow
gTTS
flask-co